Fixing deleteFile and cleaning
In MyRouteBuilder:configure():from("direct:deleteFile") I had to remove
the line
  .setHeader("fileName",
     simple("src/main/resources/files/${headers.fileName}"))
because this would not work in a jar, if the resource file is extracted
from the jar.

This has lead to deleteFile test to fail.
I have fixed this as follows:
- define folderlist
  .setHeader("folderList",simple("files,
     src/main/resources/files"))
- in FileUtilBeans search through the folderList
- throw an exeption, if the file is not found in any of the folders in the
  list

Additionally, I have cleaned the repository from keywords "velocity" and
"templates". Those have been present because this repository is a
spin-off of https://github.com/oveits/velocity-temple